fix #5: Handle unknown options
Handle unknown cli options and missing value for non boolean options.

NOTE: A missing value would be detected on the last CLI option or
command. For options or commands in the middle the executed shell line,
the parser tries to use the next element in the argv table (be it an
actual value or another cli option).